+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Renamed `AssetViews` to `Gallery`
 - Renamed `SectionViews` to `Sections`
 - Renamed `MetaTags::register()` to `MetaTags::widget()` to match the other widgets
+- Moved `Bootstrap` class to base package namespace for consistency
 
 ## 2.0.1 (Nov 4, 2023)